Feature: Soy Views Renderer
This changeset introduces a new Micronaut package to render views
via *[Soy](https://github.com/google/closure-templates)* (also
known as *Closure Templates*).

For now, the renderer simply uses `SoyTofu` to render views from
Soy template sources. There are more efficient ways to render Soy
templates in Java, and I'm planning to add those, too, subject to
feedback from the Micronaut team.

I've based this proposal on the structure of `views-freemarker`,
to make sure I've adopted Micronaut's coding style as closely as
I can. There is a corresponding set of unit tests as well, that
pass, at least locally.

_NOTE:_ This change requires a patched version of Soy Java, which
applies a PR upstream to support querying a `SoyTofu` instance for
templates (see: google/closure-templates#185). A patched JAR for
Soy can be installed with the script `patch-soy.sh` added at the
root of the repo. This is run in Travis, for instance, but can be
eliminated when the upstream PR is merged and a release is cut.

Changes made:
- [x] Support for `SoyTofu` - i.e. rendering from template sources
  - [x] `SoyTofuRenderer` - where the rendering happens
  - [x] `SoyTofuViewsRendererConfiguration` - where a `SoyFileSet`
        is loaded and configured
  - [x] `SoyFileSetProvider` - uses DI to resolve a `SoyFileSet`,
        which is used to back the `SoyTofu.Renderer`
- [x] Unit tests for `SoyTofuViewsRenderer`
- [ ] Coming shortly: `SoySauce` - i.e. rendering via templates
      pre-compiled to Java sources