fieldset templates TCK
This PR adds a Technology compatibility kit for the rendering of views with templates. It adds an implementation with thymeleaf and a partial implementation with Apache freemaker.

We could generate these templates when the user selects Thymleaf in Micronaut  Starter and run the TCK suite in a starter functional test to ensure the templates behave as we expect.